remove minor inconsistencies in naming and add parse function for vl53l4cd sensor

+++ src/vl6180x/vl6180x.h
@@ -36,7 +36,7 @@ extern "C"
      */
     typedef struct
     {
-        uint32_t distance_mm;
+        uint8_t distance_mm;
     } vl6180x_sample_data_t;
 
     /**
